Separate disco handling from backend decision
