+++ src/satosa/micro_services/custom_routing.py
@@ -7,20 +7,17 @@ from satosa.internal import InternalData
 from .base import RequestMicroService
 from ..exception import SATOSAConfigurationError
 from ..exception import SATOSAError
-from ..exception import SATOSAStateError
 
 
 logger = logging.getLogger(__name__)
 
 
 class CustomRoutingError(SATOSAError):
-    """
-    SATOSA exception raised by CustomRouting rules
-    """
+    """SATOSA exception raised by CustomRouting rules"""
     pass
 
 
-class DecideBackendByTargetIdP(RequestMicroService):
+class DecideBackendByTargetIssuer(RequestMicroService):
     """
     Select target backend based on the target issuer.
     """
@@ -38,14 +35,11 @@ class DecideBackendByTargetIdP(RequestMicroService):
         self.default_backend = config['default_backend']
 
     def process(self, context:Context, data:InternalData):
-        """
-        Set context.target_backend based on the target issuer (context.target_entity_id)
+        """Set context.target_backend based on the target issuer"""
 
-        :param context: request context
-        :param data: the internal request
-        """
         target_issuer = context.get_decoration(Context.KEY_TARGET_ENTITYID)
         if not target_issuer:
+            logger.info('skipping backend decision because no target_issuer was found')
             return super().process(context, data)
 
         target_backend = (
@@ -62,45 +56,6 @@ class DecideBackendByTargetIdP(RequestMicroService):
 
         context.target_backend = target_backend
         return super().process(context, data)
-
-
-class DecideBackendByDiscoIdP(DecideBackendByTargetIdP):
-    def __init__(self, config:dict, *args, **kwargs):
-        super().__init__(config, *args, **kwargs)
-
-        self.disco_endpoints = config['disco_endpoints']
-        if not isinstance(self.disco_endpoints, list):
-            raise CustomRoutingError('disco_endpoints must be a list of str')
-
-    def register_endpoints(self):
-        """
-        URL mapping of additional endpoints this micro service needs to register for callbacks.
-
-        Example of a mapping from the url path '/callback' to the callback() method of a micro service:
-            reg_endp = [
-                ("^/callback1$", self.callback),
-            ]
-
-        :rtype List[Tuple[str, Callable[[satosa.context.Context, Any], satosa.response.Response]]]
-
-        :return: A list with functions and args bound to a specific endpoint url,
-                 [(regexp, Callable[[satosa.context.Context], satosa.response.Response]), ...]
-        """
-
-        return [
-            (path , self._handle_disco_response)
-            for path in self.disco_endpoints
-        ]
-
-    def _handle_disco_response(self, context:Context):
-        target_issuer_from_disco = context.request.get('entityID')
-        if not target_issuer_from_disco:
-            raise CustomRoutingError('no valid entity_id in the disco response')
-
-        context.decorate(Context.KEY_TARGET_ENTITYID, target_issuer_from_disco)
-        data_serialized = context.state.get(self.name, {}).get('internal', {})
-        data = InternalData.from_dict(data_serialized)
-        return self.process(context, data)
 
 
 class DecideBackendByRequester(RequestMicroService):
